Better dependency tracking for integration tests
Summary:
Now that we split integration tests into multiple "test" dune rules, we can also properly attach the dependency of each *.py, *.pysa, *.models file to the right test.
The advantage is that, if we update a .pysa file, and run `make test`, only the corresponding test will be ran.

Reviewed By: tianhan0

Differential Revision: D82020944

fbshipit-source-id: 446460b02aca92b46a66e168bfe907c39d95c0b3