Use FunctionParameter and FunctionSignature in model query execution
Summary:
# Context
Right now, to execute model queries, we use information inferred directly from the AST of functions and methods.
Most notably, we use the type annotation parsed from the AST. The transformation from an expression to a proper `Type.t` happens in `PyrePysaEnvironment.ReadOnly.parse_annotation`.

# Problem
This will be a problem for pyrefly, as we obviously can't parse an annotation into a `Type.t`. Instead, we need to get the information from pyrefly, and get a `PysaType.t`.

# Solution
Instead of using the raw AST during model query execution, let's use `FunctionParameter.t` and `FunctionSignature.t` defined in the `PyrePysaApi`. This is a common representation between pyre1 and pyrefly. In a follow-up, we can hook up the information from pyrefly to run model generation, without having to parse types during model generation.

Reviewed By: tianhan0

Differential Revision: D82286520

fbshipit-source-id: 5f7bc6337fb94524940ef236048f6f7c861915c6