Add new internal config option `disable_type_errors_in_ide`
Summary:
Having a way to enable/disable IDE squiggles in Pyre1 will be helpful for syncing Pyre1->Pyrefly migration across both CLI/CI and IDE. With this option, we'll be able to switch off Pyre1 IDE squiggle and switch on Pyrefly IDE squiggle all within the same diff.

This diff adds the config option to Pyre1 internal client (since the migration is not relevant outside of IG, we don't need it in the OSS client). Next diff will teach the Pyre1 language server to read from this config.

Reviewed By: kinto0

Differential Revision: D82262522

fbshipit-source-id: 0ca61679e69e42374d1a07b98b3a6548487f58a1