Fix diagnostic code being set to `pyre (documentation link)`
Summary:
While analyzing some python errors for devmate, I saw that a lot of error codes were set to `pyre (documentation link)`. I traced it back to this line:
`        code=None if code_description is None else "pyre (documentation link)",
` in type_error_handler.py.

I'm not sure if this was intentional because "pyre (documentation link)" is not a real diagnostic code. The

See query: https://fburl.com/daiquery/idoh43zn

Submitting this diff to see if this change is acceptable.

{F1982146373}

Reviewed By: maggiemoss

Differential Revision: D82871788

fbshipit-source-id: b5371a661488f36a0200055435ef1d01578fe1f6