Store class decorators in shared memory for model queries
Summary:
In order to handle class decorator constraints in our model queries, we need to AST of decorators for each class when using pyrefly.

In this diff, we add a step that matches the information from pyrefly about classes with the AST, and build a mapping of fully qualified class names to decorator expressions.

Reviewed By: tianhan0

Differential Revision: D84073426

fbshipit-source-id: 9e17272f6dcf55c1949b45ee93c3442f492906ac