Move parsing of types of expressions in a later step
Summary:
When using pyrefly as source of types, we don't need the map of expression location to types until later in the analysis. Therefore, let's move parsing of types of expressions later in the analysis.
This way, users can run model verification and model queries without parsing types of expressions.

Reviewed By: tianhan0

Differential Revision: D85139799

fbshipit-source-id: f00d69de1c246906ef0b1e0768c0d25705232395