Store callables without a def statement in Target.CallablesSharedMemory
Summary:
As titled.

We are not currently storing callables that don't have an associated "def .." statement in `Target.CallablesSharedMemory`. This is possible when using pyrefly, for synthesized callables (for instance, `__init__` of a dataclass).

This diff stores those targets in `Target.CallablesSharedMemory` and uses information from pyrefly for all the fields whenever possible. For fields depending on the Ast, we use `AstResult.t`.

Reviewed By: tianhan0

Differential Revision: D85152185

fbshipit-source-id: 41af4aeb24223a9d0148b143d0af4e38d5d7753c