Move call graph building logic in a separate module
Summary:
# Problem
We want to parse call graphs from the pyrefly results.
However, the call graph definition in OCaml is in the same module as the logic that builds the call graph and depends on `PyrePysaApi` which depends on `PyreflyApi` indirectly.
Due to this, using `CallGraph` in `PyreflyApi` would create a cyclic dependency.

# Solution
Split `CallGraph` into two modules:
* `CallGraph` contains the reprentation of call graphs
* `CallGraphBuilder` contains the logic to build call graphs from the type checking state

Reviewed By: tianhan0

Differential Revision: D85261877

fbshipit-source-id: 77d9799726da5a5e4560332fbf5b5e4e9b381bba