Use if_called in pysa call graph
Summary:
We changed the representation of call graph edges for Name and AttributeAccess nodes in Pyrefly.
This was motivated by the fact that a Name ast node can refer to a class, and hence its callees would be the `__new__` and `__init__`.
Let's change the representation in Pysa to match the Pyrefly representation.

Reviewed By: tianhan0

Differential Revision: D85686285

fbshipit-source-id: b24f2812742d55f117fc23bb752d635680179b97