Fix location of define callees in the call graph
Summary:
# Problem
In Pyre, the location of a `def ..` statement does NOT include its decorators, while in Pyrefly, it does.
That leads to inconsistencies for `DefineCallees` edges: they have different locations between Pyre and Pyrefly.

# Solution
Let's use a consistent location between Pyre and Pyrefly. It doesn't seem possible to get the location of the define without the decorators in Pyrefly, so let's use the Pyrefly location as the source of truth.

Reviewed By: tianhan0

Differential Revision: D85775947

fbshipit-source-id: 05ec404dbe3e12680da2251df377597081dbfcd6