Don't qualify the original function name in decorated bodies
Summary:
To handle decorators, Pysa creates artificial functions that corresponds to applying all the decorators on the decorated function, i.e `decorator1(decorator2(foo))`.
When using Pyre1, the original function is represented as a fully qualified symbol, therefore it needs to be an attribute access expression: `a.b.foo`.
When using Pyrefly, this is not really necessary. Let's just use an identifier expression.

Reviewed By: tianhan0

Differential Revision: D87546497

fbshipit-source-id: 1d335ae126ad6b6246a8e26dc0daa853746f63ad