Use types from pyrefly during the taint analysis
Summary:
During the taint analysis, we sometimes use the type of an expression. The most common case use case is to add `via:scalar` breadcrumbs. Another use case is for `via-type:<type>` breadcrumb.

This diff implements support for this when using Pyrefly. Pyrefly exports a mapping from location to type for each module, and parse it in Pysa.
This diff updates the logic to pull information from the parsed result.

Reviewed By: tianhan0

Differential Revision: D87784300

fbshipit-source-id: e73e1b3a2a4535e9f583bee8b2d5165e1f3ae79f