Add module qualifier and define name to InContext API
Summary:
The `InContext` API is always created within a module and function.
The `TypeOfExpressionSharedMemory` API requires a module name (when used with pyrefly) and a define name (when used with pyre1).
Right now it requires to pass the define name as an argument, and retrieves the module name from the define name.
This is unnecessary since the `pyre_in_context` object could just store information about the current module and function.

This is what this diff does: store the module and define name in the `InContext` object, so it can be used by the `TypeOfExpressionsSharedMemory`.

Reviewed By: tianhan0

Differential Revision: D87925721

fbshipit-source-id: b3b1186af73a8a5353cbd30d0e01dde40eef4551