Remove the use of `qualified_name_of_define`
Summary:
When using pyrefly, `qualified_name_of_define` does not give the correct define name, since it assumes the define name is qualified in the AST. However, we disable qualification when using pyrefly.

Let's get rid of most of the uses of `qualified_name_of_define`. In most cases, we can get the define name from the target.

Reviewed By: tianhan0

Differential Revision: D87929862

fbshipit-source-id: a33134c0727f8715166e1a89bb6234693e68542b