Fix define name when exporting issues as errors
Summary:
When pysa runs without `--save-results-to`, issues are printed directly to stdout as JSON.
Those errors have a field `"define"` which is supposed to be the fully qualified name of the function where the issue is.
When using pyrefly, this is just the function name because it relied on the qualification step in Pyre1.
Let's fix it.

Reviewed By: tianhan0

Differential Revision: D87933338

fbshipit-source-id: 23f1994df308496be2976e74784825428329b692