Fix override skipping when using pyrefly
Summary:
# Problem
Skipping overrides using `SkipOverrides` on a builtin class like `MutableMapping` doesn't work because we expand the overrides directly in pyrefly before we can skip it.

# Solution
To fix this, we export a special target `OverrideSubset` which we can treat correctly on the pysa side. When we see a call graph edge to a `OverrideSubset` with a base method that is `SkipOverrides`, we discard the override subset, and keep only the base method.

Reviewed By: tianhan0

Differential Revision: D88070296

fbshipit-source-id: e672278e346798628d3e29bb585d141a03ac03e8