Move pysa test helpers in a pysa module
Summary:
# Problem
In our integration tests, we use a set of helpers like `_taint_source`, `_taint_sink` and `_tito`.
Those are currently injected in the builtins module.

This is a bit complicated to replicate with pyrefly, which has its own bundled typeshed.

# Solution
Let's move all these functions into a `pysa/__init__.pyi` module that we inject in the test environment.

Reviewed By: tianhan0

Differential Revision: D88378794

fbshipit-source-id: b71efd4ba283d62331b679205e8363d1515086d3