Rename NotInEnvironment into BaseModuleNotInEnvironment
Summary:
To make it clearer that this error is only used when the base module of a user-provided name is not in the environment.

With pyrefly, this becomes a common error since it only type checks what is transitively imported from the roots. This means we end up with models for third party libraries generating a lot of errors. We will ned to deal with it in a smart way.

Reviewed By: yangdanny97

Differential Revision: D88383573

fbshipit-source-id: 48d50f98805d60c3e31b92514d3ac501fafd541e