Don't normalize assertions when using pyrefly
Summary:
# Problem
When we build the control flow graph for a function, we normalize conditions in `if/elif/else` statements. I believe this is done to keep our pattern matching logic for type refineming simple (e.g, `if isinstance(a, int)` will refine the type of a).

However, these creates expressions that don't exist in the original AST.
When using pyrefly as a frontend for Pysa, these expressions won't have call graph edges, leading to false negatives.

# Solution
Simply disable the normalization of assertions when using pyrefly.

Reviewed By: tianhan0

Differential Revision: D88753635

fbshipit-source-id: 743518fee392fc2ca121dfc1414c3fc07397248d