Run most end-to-end integration tests with pyrefly
Summary:
This diff allows use to use pyrefly as a frontend for most of our end-to-end integration tests.
Since end-to-end integration tests aren't supposed to have type errors, we also have to fix a bunch of them.
Note that we still need to verify that the output is what we expect, for each test. That will be done later.

Reviewed By: tianhan0

Differential Revision: D89056682

fbshipit-source-id: 4e9cd94d83853b17f50ef853f65afcadee3191f7